Move app related files into App folder
