+++ src/app/App.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render } from '@testing-library/react'
-import App from 'App'
+import App from 'app/App'
 
 test('renders Hello World!', () => {
   const { getByText } = render(<App />)
